Merge branch: Fix merge HTTP headers with cURL options (#63)
HTTP headers were not merged properly, and were also not accounting for possible casing differences.
Follow-up of https://github.com/simplepie/simplepie/pull/912
It was for instance not possible to override the `Accept` header from the cURL options.
The existing code could not work properly because we were merging arrays of strings and not arrays of key-values, leading to strange duplicates.

Downstream:
* https://github.com/FreshRSS/FreshRSS/pull/8246

Upstream:
* https://github.com/simplepie/simplepie/pull/956